remove redundant tests
